hyungjup: check_valid_map 구현

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: hyungjup <hyungjup@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 20:01:51 by hyungjup          #+#    #+#             */
-/*   Updated: 2022/11/09 11:44:21 by hyungjup         ###   ########.fr       */
+/*   Updated: 2023/09/05 16:41:09 by hyungjup         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
 	size_t	i;
